fix(route): update paths for Transactions
